reafactor code and folders and before refactoring of namespace
